+++ templates/contact.HTML
@@ -49,8 +49,7 @@
         </form>
         <ul class="nav navbar-nav navbar-right">
           <li><a href="{% url 'index_itbursa' %}">главная</a></li>
-          <li><A href="{% url 'index_itbursa' %}">преподаватели</A></li>
-          <li><A href="{% url 'student_list' %}">студенты</A></li>
+          <li><A href="{% url 'students:student_list' %}">студенты</A></li>
           <li><a href="#">контакты</a></li>
         </ul>
       </div>
