+++ templates/index.HTML
@@ -51,7 +51,7 @@
           </form>
           <ul class="nav navbar-nav navbar-right">
             <li><a href="#">главная</a></li>
-            <li><a href="{% url 'student_list' %}">студенты</a></li>
+            <li><a href="{% url 'students:student_list' %}">студенты</a></li>
             <li><a href="{% url 'contacts_itbursa' %}">контакты</a></li>
           </ul>
         </div>
@@ -63,50 +63,32 @@
       <div class="row">
         <div class="col-sm-12 col-md-12">
           <div class="thumbnail">
-            <img src="{% static "IMG/slider1.jpg"%}" alt="...">
+            <img src="{% static "IMG/slider1.jpg" %}" alt="...">
 
           </div>
         </div>
       </div>
     </div>
+
     <!--Курс JS -->
     <div class="row">
+      {% for course in courses %}
         <div class="col-sm-4 col-md-4" >
           <div class="thumbnail" id = 'python'>
-            <a href="/courses/{{courses.0.id}}/">
-            <img src="{% static "IMG/Python_course.jpg"%}" alt="...">
+            <a href="{% url 'courses:courses' course.id %}">
+
+              <img src = "/static/IMG/Course_{{course.id}}.jpg" alt="...">
               <div class="caption">
-                <h3>{{courses.0.info}}</h3></a>
+                <h3>{{course.info}}</h3></a>
                 <br>
-                 <p>{{courses.0.discription}}</p>
+                 <p>{{course.discription}}</p>
                   <p><a href="http://pybursa.com/" class="btn btn-success btn-lg btn-block" role="button">ЗАПИСАТЬСЯ</a> </p>
               </div>
-          </div>
+            </div>
         </div>
-        <!--Курс JS -->
-        <div class="col-sm-4 col-md-4" >
-          <div class="thumbnail" id = 'django' >
-            <a href="/courses/{{courses.1.id}}/">
-            <img src="{% static "IMG/Django_course.jpg"%}" alt="...">
-              <div class="caption">
-                <h3>{{courses.1.info}}</h3></a>
-                <p>{{courses.1.discription}}</p>
-                  <p><a href="http://pybursa.com/" class="btn btn-success btn-lg btn-block" role="button">ЗАПИСАТЬСЯ</a> </p>
-              </div>
-          </div>
-        </div>
-        <div class="col-sm-4 col-md-4" >
-          <div class="thumbnail" id = 'django' >
-            <a href="/courses/{{courses.2.id}}/">
-            <img src="{% static "IMG/JS_course.jpg"%}" alt="...">
-              <div class="caption">
-                <h3>{{courses.2.info}}</h3></a>
-                <p>{{courses.2.discription}}</p>
-                  <p><a href="http://pybursa.com/" class="btn btn-success btn-lg btn-block" role="button">ЗАПИСАТЬСЯ</a> </p>
-              </div>
-          </div>
-        </div>
+      {% endfor %}
     </div>
+
 
     <!--Футер -->
     <div class = 'row'>
